Create pure3.md
This is very much a work in progress and at this point is not even complete. I crafted this from what had been posted in Standard PR #231 and Standard Issue #227. I frankly was not sure what we should do in terms of how we release the information or a patch once a vulnerability is confirmed. 

@hexsecs and others, please add on.