additions to PURE3
I fleshed out some proposed procedures. We need reviews to see if what I wrote is feasible. 

I am not sure what would constitute the "Security Analysis" piece for this PURE. Would we need to detail how we would determine the extent of the security risk?

I am assuming that the "backwards compatibility" element would be tied to what the patches might change and not anything to do with the procedure. If so, then backwards compatibility would vary with the vulnerability and its remediation.